+++ src/raport_slotow/views/ewaluacja.py
@@ -3,21 +3,20 @@ from urllib.parse import urlencode
 
 from django.http import HttpResponseRedirect
 from django.urls import reverse
+from django.utils import timezone
 from django.views.generic import FormView
 from django_filters.views import FilterView
 from django_tables2 import SingleTableMixin
-
 from formdefaults.helpers import FormDefaultsMixin
+
+from django_bpp.version import VERSION
 from nowe_raporty.views import BaseRaportAuthMixin
 from raport_slotow.filters import RaportSlotowUczelniaEwaluacjaFilter
 from raport_slotow.forms.ewaluacja import ParametryRaportSlotowEwaluacjaForm
 from raport_slotow.models import RaportUczelniaEwaluacjaView
 from raport_slotow.tables import RaportSlotowEwaluacjaTable
+from raport_slotow.uczelnia_helper import uczelnia_dla_odczytu
 from raport_slotow.util import MyExportMixin
-
-from django.utils import timezone
-
-from django_bpp.version import VERSION
 
 
 class ParametryRaportSlotowEwaluacja(
@@ -56,6 +55,18 @@ class RaportSlotowEwaluacja(
 
     def get_form(self, dct):
         return ParametryRaportSlotowEwaluacjaForm(dct)
+
+    def get_table_kwargs(self):
+        kwargs = super().get_table_kwargs()
+        uczelnia = uczelnia_dla_odczytu(self.request)
+        if not (uczelnia and uczelnia.uzywaj_wydzialow):
+            exclude = tuple(kwargs.get("exclude", ()))
+            kwargs["exclude"] = (
+                *exclude,
+                "aktualny_wydzial",
+                "afiliowany_wydzial",
+            )
+        return kwargs
 
     def get(self, request, *args, **kw):
         form = self.get_form(request.GET)
@@ -103,12 +114,12 @@ class RaportSlotowEwaluacja(
                 "autorzy__autor",
                 "autorzy__dyscyplina_naukowa",
                 "autorzy__autor__tytul",
+                "autorzy__autor__aktualna_jednostka__wydzial",
+                "autorzy__jednostka__wydzial",
                 "autor_dyscyplina__dyscyplina_naukowa",
                 "autor_dyscyplina__subdyscyplina_naukowa",
             )
             .prefetch_related(
-                "autorzy__autor__aktualna_jednostka",
-                "autorzy__jednostka",
                 "rekord__zrodlo__punktacja_zrodla_set",
             )
             .only(
@@ -148,8 +159,12 @@ class RaportSlotowEwaluacja(
                 "autorzy__profil_orcid",
                 "autorzy__jednostka_id",
                 "autorzy__jednostka__nazwa",
+                "autorzy__jednostka__wydzial_id",
+                "autorzy__jednostka__wydzial__nazwa",
                 "autorzy__autor__aktualna_jednostka_id",
                 "autorzy__autor__aktualna_jednostka__nazwa",
+                "autorzy__autor__aktualna_jednostka__wydzial_id",
+                "autorzy__autor__aktualna_jednostka__wydzial__nazwa",
                 "autor_dyscyplina__dyscyplina_naukowa",
                 "autor_dyscyplina__dyscyplina_naukowa__id",
                 "autor_dyscyplina__dyscyplina_naukowa__nazwa",